requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Gradient boosting models
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Hyperparameter tuning
optuna>=3.4.0

# Experiment tracking
mlflow>=2.9.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Monitoring
evidently>=0.4.0

# Feature engineering
feature-engine>=1.6.0

# Visualization (EDA)
matplotlib>=3.8.0
seaborn>=0.13.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.27.0

# Imbalanced learning
imbalanced-learn>=0.11.0

# Serialization
joblib>=1.3.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
